Move enum IssuerState to protocols layer
