+++ aries_vcx/tests/utils/devsetup_agent.rs
@@ -6,7 +6,7 @@ pub mod test {
     use aries_vcx::handlers::connection::public_agent::PublicAgent;
     use aries_vcx::handlers::issuance::holder::Holder;
     use aries_vcx::handlers::issuance::holder::test_utils::get_credential_offer_messages;
-    use aries_vcx::handlers::issuance::issuer::{Issuer, IssuerState};
+    use aries_vcx::handlers::issuance::issuer::Issuer;
     use aries_vcx::handlers::proof_presentation::prover::{Prover, ProverState};
     use aries_vcx::handlers::proof_presentation::prover::test_utils::get_proof_request_messages;
     use aries_vcx::handlers::proof_presentation::verifier::{Verifier, VerifierState};
@@ -25,6 +25,7 @@ pub mod test {
     use aries_vcx::protocols::connection::invitee::state_machine::InviteeState;
     use aries_vcx::protocols::connection::inviter::state_machine::InviterState;
     use aries_vcx::protocols::issuance::holder::state_machine::HolderState;
+    use aries_vcx::protocols::issuance::issuer::state_machine::IssuerState;
     use aries_vcx::settings;
     use aries_vcx::utils::devsetup::*;
     use aries_vcx::utils::provision::{AgencyClientConfig, AgentProvisionConfig, provision_cloud_agent};
